Create Bulk Synthetic Data

Creates a series of waiting lists, one for each row in the dataframe parameter and joins them together into one dataframe with relevant creation criteria

Usage

create_bulk_synthetic_data(bulk_data)

Arguments

bulk_data

A dataframe object, each row being a waiting list with parameters to generate the synthetic data. A sample data.frame is available as demo-data

Value

Dataframe of waiting lists for each specified site and specialty, opcs etc

Examples

create_bulk_synthetic_data(demo_df)
#> # A tibble: 67,685 × 8
#>    hospital_site main_spec_code opcs4_code pat_id addition_date removal_date
#>    <chr>                  <dbl> <chr>       <int> <date>        <date>      
#>  1 ABC001                   100 T202            1 2024-01-01    2024-03-07  
#>  2 ABC001                   100 T202            2 2024-01-01    2024-01-13  
#>  3 ABC001                   100 T202            3 2024-01-01    2024-02-12  
#>  4 ABC001                   100 T202            4 2024-01-01    2024-02-16  
#>  5 ABC001                   100 T202            5 2024-01-01    2024-01-02  
#>  6 ABC001                   100 T202            6 2024-01-01    2024-01-05  
#>  7 ABC001                   100 T202            7 2024-01-01    2024-01-06  
#>  8 ABC001                   100 T202            8 2024-01-01    2024-02-06  
#>  9 ABC001                   100 T202            9 2024-01-01    2024-01-17  
#> 10 ABC001                   100 T202           10 2024-01-01    2024-01-25  
#> # ℹ 67,675 more rows
#> # ℹ 2 more variables: wait_length <dbl>, rott <lgl>