A Demo 'data.frame' Object

A pre-created data.frame ready to be used to test the create_bulk_synthetic_data and create_waiting_list functions. Each row of the data.frame represents an individual waiting list for which the site, specialty, OPCS code(s) and respective mean wait, arrival rate, start_date, sd and rott can be specified. It allows the user to see an example of the structure of the data.frame required by the create_bulk_synthetic_data function to create your synthetic dataset.

Usage

demo_df

Format

A dataframe with 5 rows and 9 columns:

hospital_site

Character. Hospital site code of the waiting list.

main_spec_code

Numeric. Main specialty code of the waiting list.

opcs4_code

Character. OPCS4 code(s) of the procedure(s) on the waiting list.

n

Numeric. Number of days for which to create synthetic waiting list data.

mean_arrival_rate

Numeric. Mean number of arrivals per day.

mean_wait

Numeric. Mean wait time for treatment/on waiting list.

start_date

Character. Date from which to start generated waiting list in format yyyy-mm-dd.

sd

Numeric. Standard deviation.

rott

Numeric. Proportion of referrals to be randomly flagged as ROTT